ImageCapitoning_To_Text/requirements.txt
streamlit
Pillow
requests
transformers
torch